refactor(core): rename condition group to expression
