+++ packages/core/src/classes/expression/base-condition.ts
