+++ packages/core/src/classes/transformer/document-client-request-transformer.ts
@@ -20,9 +20,9 @@ import {dropProp} from '../../helpers/drop-prop';
 import {getConstructorForInstance} from '../../helpers/get-constructor-for-instance';
 import {isEmptyObject} from '../../helpers/is-empty-object';
 import {parseKey} from '../../helpers/parse-key';
-import {KeyCondition} from '../condition/key-condition';
+import {KeyCondition} from '../expression/key-condition';
 import {Connection} from '../connection/connection';
-import {ExpressionBuilder} from '../expression-builder';
+import {ExpressionBuilder} from '../expression/expression-builder';
 import {EntityManagerUpdateOptions} from '../manager/entity-manager';
 import {AttributeMetadata} from '../metadata/attribute-metadata';
 import {DynamoEntitySchemaPrimaryKey} from '../metadata/entity-metadata';
